* Adding spdlog MIT license file, * db_string class that wraps C string, * finally some (but not all) of the code moved from headers to source file (some day this process will be completed), * added VS version resource, * segmente_container for any type of data, storing it in blocks of fixed size (WIP) * utils function for calculating size of an object (statedb::utils::get_object_size)

+++ StateDB/db_wrapper.h
@@ -1,7 +1,5 @@
 #pragma once
-#include "db_header.h"
 #include "db.h"
-#include "stdafx.h"
 
 namespace fs = std::filesystem;
 
@@ -41,7 +39,7 @@ namespace statedb {
 		void verify_file()
 		{
 			spdlog::debug("Verifying file...");
-			size_t size = fs::file_size(db_filepath);
+			uintmax_t size = fs::file_size(db_filepath);
 			spdlog::debug("File size = {}", size);
